requirements.txt
BeautifulSoup
requests
pandas
fpdf
seaborn
matplotlib.pyplot